+++ src/main/java/com/esgi/repository/ReviewRepository.java
@@ -15,5 +15,5 @@ import java.math.BigDecimal;
 public interface ReviewRepository extends JpaRepository<ReviewEntity, Long> {
 
     @Query("SELECT SUM(reviews.rating)/COUNT(reviews) FROM ReviewEntity reviews WHERE reviews.idmovie = :idmovie")
-    BigDecimal getRating(@Param("idmovie") int idmovie);
+    BigDecimal getRating(@Param("idmovie") Long idmovie);
 }
